+++ stories/helpers/i18n.js
@@ -1,7 +1,7 @@
 import i18n from "i18next";
 import { initReactI18next } from "react-i18next";
 
-import resources from '../locales'
+import resources from '../../locales'
 
 i18n
   .use(initReactI18next) // passes i18n down to react-i18next
